server/requirements.txt
flask
mss
pillow
